feat: Make bech32 address a link in Address.tsx + Fix computation of bech32 from hex + Refactor + Fix lint
