+++ client/src/app/components/stardust/TransactionPayload.tsx
@@ -6,13 +6,13 @@ import { NameHelper } from "../../../helpers/stardust/nameHelper";
 import { formatAmount } from "../../../helpers/stardust/valueFormatHelper";
 import NetworkContext from "../../context/NetworkContext";
 import AsyncComponent from "../AsyncComponent";
-import Bech32Address from "../chrysalis/Bech32Address";
 import FiatValue from "../FiatValue";
 import Modal from "../Modal";
 import { ModalIcon } from "../ModalProps";
 import { TransactionPayloadState } from "../TransactionPayloadState";
 import { ReactComponent as DropdownIcon } from "./../../../assets/dropdown-arrow.svg";
 import messageJSON from "./../../../assets/modals/message.json";
+import Bech32Address from "./Bech32Address";
 import "./TransactionPayload.scss";
 import Output from "./Output";
 import { TransactionPayloadProps } from "./TransactionPayloadProps";
